fix: Table cell and header renderer to be per column
