--- morph/react/get-cell-renderer.js
@@ -1,7 +1,0 @@
-export default state => {
-  return state.isTable
-    ? `const cellRenderer = ({ cellData }) => { 
-        return <${state.cell.name} value={cellData} />;
-      };`
-    : ''
-}
