--- morph/react/get-header-renderer.js
@@ -1,7 +1,0 @@
-export default state => {
-  return state.externalHeader
-    ? `const headerRenderer = ({ label }) => { 
-        return <${state.externalHeader.name} value={label} />;
-      };`
-    : ''
-}
